third commit: addition of program and milestone controller, login method in user controller

+++ src/main/java/com/visiontracker/challengeTrackerApplication/models/datamodels/CreateProgramReq.java
@@ -16,8 +16,8 @@ import java.util.List;
  */
 public class CreateProgramReq {
     private Program program;
-    private Long userId;
-    private List<Long> userIds;
+    private Integer userId;
+    private List<Integer> userIds;
     
     private String startDate;
     private String targetCompletionDate;
@@ -25,7 +25,7 @@ public class CreateProgramReq {
     public CreateProgramReq() {
     }
 
-    public CreateProgramReq(Program program, Long userId, List<Long> userIds, String startDate, String targetCompletionDate) {
+    public CreateProgramReq(Program program, Integer userId, List<Integer> userIds, String startDate, String targetCompletionDate) {
         this.program = program;
         this.userId = userId;
         this.userIds = userIds;
@@ -50,28 +50,28 @@ public class CreateProgramReq {
     /**
      * @return the userId
      */
-    public Long getUserId() {
+    public Integer getUserId() {
         return userId;
     }
 
     /**
      * @param userId the userId to set
      */
-    public void setUserId(Long userId) {
+    public void setUserId(Integer userId) {
         this.userId = userId;
     }
 
     /**
      * @return the userIds
      */
-    public List<Long> getUserIds() {
+    public List<Integer> getUserIds() {
         return userIds;
     }
 
     /**
      * @param userIds the userIds to set
      */
-    public void setUserIds(List<Long> userIds) {
+    public void setUserIds(List<Integer> userIds) {
         this.userIds = userIds;
     }
